backend-jumbo-jcldev/CencosudAdapters/OrdersAdapter/src/main/java/com/cencosud/mobile/order/model/ItemDetailR2.java
package com.cencosud.mobile.order.model;

import java.util.List;

public class ItemDetailR2{

	private String productId;
	private String description;
	private String image;
	private String price;
	private int quantity;
	private String productReference;
	private String brand;
	private String measurementUnit;
	private String skuId;
	private String addToCartLink;
	private List<Object> skuData;
	private String sellerId;

	public String getProductReference() {
		return productReference;
	}

	public void setProductReference(String productReference) {
		this.productReference = productReference;
	}

	public String getBrand() {
		return brand;
	}

	public void setBrand(String brand) {
		this.brand = brand;
	}

	public String getMeasurementUnit() {
		return measurementUnit;
	}

	public void setMeasurementUnit(String measurementUnit) {
		this.measurementUnit = measurementUnit;
	}

	public String getProductId() {
		return productId;
	}

	public void setProductId(String productId) {
		this.productId = productId;
	}

	public String getDescription() {
		return description;
	}

	public void setDescription(String description) {
		this.description = description;
	}

	public String getImage() {
		return image;
	}

	public void setImage(String image) {
		this.image = image;
	}

	public String getPrice() {
		return price;
	}

	public void setPrice(String price) {
		this.price = price;
	}

	public int getQuantity() {
		return quantity;
	}

	public void setQuantity(int quantity) {
		this.quantity = quantity;
	}

	public String getSkuId() {
		return skuId;
	}

	public void setSkuId(String skuId) {
		this.skuId = skuId;
	}

	public String getAddToCartLink() {
		return addToCartLink;
	}

	public void setAddToCartLink(String addToCartLink) {
		this.addToCartLink = addToCartLink;
	}

	public List<Object> getSkuData() {
		return skuData;
	}

	public void setSkuData(List<Object> skuData) {
		this.skuData = skuData;
	}

	public String getSellerId() {
		return sellerId;
	}

	public void setSellerId(String sellerId) {
		this.sellerId = sellerId;
	}
}
